+++ scripts/FriendlierFire/non-player.lua
@@ -1,4 +1,6 @@
 local storage = require('openmw.storage')
+local self = require("openmw.self")
+local I = require("openmw.interfaces")
 
 require("scripts.FriendlierFire.actorCommon")
 require("scripts.FriendlierFire.logic.ai")
@@ -9,11 +11,26 @@ local function onUpdate()
     if sectionFtP:get("disableSpells") then SpellUpdate() end
 end
 
+local function startAIPackage(pkg)
+    TargetChanged(pkg)
+end
+
+local function checkTargetsLeader(data)
+    local state = I.FollowerDetectionUtil.getState()
+    if state.followsPlayer then
+        data.sender:sendEvent("StopAttackingOtherFollower", { sender = self })
+    end
+end
+
 return {
     -- engineHandlers = {
     --     onUpdate = onUpdate,
     -- },
     eventHandlers = {
-        StopAttackingLeader = StopAttackingLeader,
+        StartAIPackage = startAIPackage,
+        TargetChanged = TargetChanged,
+        -- for follower -> follower aggro prevention
+        CheckTargetsLeader = checkTargetsLeader,
+        StopAttackingOtherFollower = StopAttackingOtherFollower,
     }
 }