+++ scripts/FriendlierFire/player.lua
@@ -1,7 +1,13 @@
 local storage = require('openmw.storage')
 local self = require("openmw.self")
+local I = require("openmw.interfaces")
 
 require("scripts.FriendlierFire.actorCommon")
+require("scripts.FriendlierFire.utils.dependencies")
+
+CheckDependencies(self, {
+    ["FollowerDetectionUtils.omwscripts"] = I.FollowerDetectionUtils == nil
+})
 
 local sectionPtF = storage.globalSection('SettingsFriendlierFire_playerToFollowers')
 
@@ -10,7 +16,7 @@ local function onUpdate()
 end
 
 local function localEnemyTargetsChanged(data)
-    data.actor:sendEvent("StopAttackingLeader", { target = self })
+    data.actor:sendEvent("TargetChanged", { target = self })
 end
 
 return {