+++ scripts/FriendlierFire/utils/dependencies.lua
@@ -3,6 +3,12 @@ local core = require("openmw.core")
 local l10n = core.l10n("FriendlierFire")
 local mainMod = "Friendlier Fire"
 
+---@param player any
+---@param dependencies table { key = file_name, value = boolean indicating whether required interface is missing }
+-- e.g. { ["Impact Effects.omwscripts"] = I.impactEffects == nil }  
+-- 
+-- if mod has no interfaces, set it's value to false  
+-- e.g. { ["Some Mod.omwscripts"] = false }
 function CheckDependencies(player, dependencies)
     for fileName, interfaceMissing in pairs(dependencies) do
         local filePresent = core.contentFiles.has(string.lower(fileName))