+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Friendlier Fire (OpenMW)
 
+## 2.0
+
+### Features
+
+- Incorporation of Follower Detection Util for (I shit you not) follower detection
+- Followers no longer aggro onto each other [NEEDS TESTING]
+
 ## 1.0.3
 
 ### Fixes